extractors/scrapy/pyproject.toml
[project]
name = "culture-extractor-scrapy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "newnewid>=0.5.0",
    "psycopg[binary]>=3.2.5",
    "python-dotenv>=1.1.1",
    "scrapy>=2.13.3",
    "sqlalchemy>=2.0.43",
    "culture-libraries",
]

[tool.uv.sources]
culture-libraries = { workspace = true }

[dependency-groups]
dev = [
    "ruff>=0.14.4",
]

[tool.ruff]
# Target Python 3.14+
target-version = "py314"

# Line length
line-length = 100

# Exclude common directories
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "*.egg-info",
    ".pytest_cache",
    ".ruff_cache",
]

[tool.ruff.lint]
# Enable specific rule sets
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
    "S",   # flake8-bandit (security)
]

# Ignore specific rules
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "S101",   # use of assert detected (common in scrapy)
    "ARG002", # unused method argument (common in scrapy callbacks)
]

# Allow autofix for all enabled rules
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.per-file-ignores]
# Ignore security checks in test files
"test_*.py" = ["S"]
"**/tests/*" = ["S"]

[tool.ruff.format]
# Use double quotes for strings
quote-style = "double"

# Indent with spaces
indent-style = "space"

# Unix-style line endings
line-ending = "lf"
